auth.HasValidResourceToken/auth.Middleware changes
The behavior of auth.Middleware has changed. Where previously
it would not proceed if there was not a valid account associated with
a JWT, it now sets an error and proceeds instead. This permits the
existence of tokens not associated with a currently-existing user
account. auth.HasRole still handles this error; it will return an error
if the sub claim does not correspond to an existing account.

There is now a new AuthCheck generator, auth.HasValidResourceToken,
that checks tokens not associated with a user account. See its
documentation for details.